+++ cookbooks/chub-buyspace/recipes/graphicsmagick.rb
@@ -20,6 +20,8 @@
 # Installs our custom-compiled GraphicsMagick package.
 # GraphicsMagick is a dependency of BuySpace
 
+include_recipe 'chub-buyspace::default'
+
 %w{
 	libpng++-dev
 	libjpeg-dev
@@ -29,7 +31,6 @@
 		action :install
 	end
 end
-
 
 remote_file "/var/cache/apt/archives/graphicsmagick-1.3.18_amd64.deb" do
 	source "http://dev4:8000/packages/graphicsmagick-1.3.18_amd64.deb"