+++ cookbooks/chub-castle/roles/castle.rb
@@ -1,2 +1,0 @@
-name "castle"
-run_list "recipe[chub_sensu_sso::client]", "recipe[chub-castle::default]"