+++ cookbooks/chub_castle/attributes/default.rb
@@ -1,0 +1,6 @@
+default[:chub_castle][:app_url] = "http://mpbamboo.nexus.commercehub.com/browse/SSO-CAS/latestSuccessful/artifact/shared/castle.war/castle.war"
+default[:chub_castle][:java_heap_size] = "512M"
+default[:chub_castle][:keystore_file] = "keystore.jks"
+default[:chub_castle][:truststore_file] = "truststore.jks"
+default[:chub_castle][:keystore_password] = "changeit"
+default[:chub_castle][:truststore_password] = "changeit"