+++ cookbooks/chub_java/recipes/oracle7.rb
@@ -1,0 +1,4 @@
+node.default[:java][:install_flavor] = 'oracle'
+node.default[:java][:jdk_version] = '7'
+
+include_recipe 'java::oracle'