+++ roles/vault.rb
@@ -1,3 +1,3 @@
 name "vault"
 description "A node for the Vault service"
-run_list "recipe[base]", "recipe[chub_sensu_sso::client]", "recipe[chub-vault]"
+run_list "recipe[base]", "recipe[chub_sensu_sso::client]", "recipe[chub_vault]"